app/requirements.txt
flask==3.0.0
requests==2.31.0
discord.py==2.3.2
python-dotenv==1.0.0
paramiko==3.4.0
audioop-lts
